feat: remove guardian timelock and add missing SDK functions (#14)
Remove the guardian timelock mechanism that provided no real security:
- Simplified OneAuthRecoveryBase.sol to apply guardian config changes immediately
- Removed guardian timelock configuration from OneAuthValidator.sol and SDK
- Updated all integration tests and invariants

Add 6 missing SDK functions through Rust→WASM→TypeScript pipeline:
- Credential management: encodeAddCredential, encodeRemoveCredential
- Guardian config: encodeSetGuardianConfig
- Guardian signatures: encodeSingleGuardianSig, encodeDualGuardianSig, encodeGuardianEntries
- Added 4 new TypeScript interfaces (AddCredentialInput, SetGuardianConfigInput, NewCredential, GuardianEntry)

Complete package distribution with comprehensive JSDoc:
- Rebuilt WASM binary (189KB → 209KB) with new exports
- Added module-level documentation with 6 flow examples (install, credentials, guardian config, signing, recovery, Guardian.sol)
- Every exported function has @param and @returns documentation
- Organized re-exports by functional group in index.ts

All tests passing: 30 Rust unit + 9 golden + 105 Solidity = 144 total

+++ test/OneAuth/invariant/handlers/CredentialHandler.sol
@@ -70,7 +70,7 @@ contract CredentialHandler is Test {
         }
 
         vm.prank(actor);
-        try validator.onInstall(abi.encode(keyIds, creds, address(0), address(0), uint48(0))) {
+        try validator.onInstall(abi.encode(keyIds, creds, address(0), address(0), uint8(0))) {
             ghost_isInstalled[actor] = true;
             ghost_credentialCount[actor] = count;
             for (uint256 i; i < count; i++) {
@@ -88,7 +88,7 @@ contract CredentialHandler is Test {
         try validator.onUninstall("") {
             ghost_isInstalled[actor] = false;
             ghost_credentialCount[actor] = 0;
-            // onUninstall clears guardian/pending/timelock on-chain, sync ghost state
+            // onUninstall clears guardian on-chain, sync ghost state
             if (address(recoveryHandler) != address(0)) {
                 recoveryHandler.ghostClearGuardianState(actor);
             }
